refactor: rename DocArrayStack to DocVec
