+++ docarray/array/__init__.py
@@ -1,4 +1,4 @@
 from docarray.array.array.array import DocList
-from docarray.array.stacked.array_stacked import DocVec
+from docarray.array.doc_vec.doc_vec import DocVec
 
 __all__ = ['DocList', 'DocVec']
