+++ docarray/array/array/array.py
@@ -31,7 +31,7 @@ if TYPE_CHECKING:
     from pydantic import BaseConfig
     from pydantic.fields import ModelField
 
-    from docarray.array.stacked.array_stacked import DocVec
+    from docarray.array.doc_vec.doc_vec import DocVec
     from docarray.proto import DocumentArrayProto
     from docarray.typing import TorchTensor
     from docarray.typing.tensor.abstract_tensor import AbstractTensor
@@ -257,11 +257,11 @@ class DocList(
         """
         Convert the DocArray into a DocArrayStacked. `Self` cannot be used
         afterwards
-        :param tensor_type: Tensor Class used to wrap the stacked tensors. This is useful
+        :param tensor_type: Tensor Class used to wrap the doc_vec tensors. This is useful
         if the BaseDoc has some undefined tensor type like AnyTensor or Union of NdArray and TorchTensor
         :return: A DocArrayStacked of the same document type as self
         """
-        from docarray.array.stacked.array_stacked import DocVec
+        from docarray.array.doc_vec.doc_vec import DocVec
 
         return DocVec.__class_getitem__(self.document_type)(
             self, tensor_type=tensor_type
@@ -274,7 +274,7 @@ class DocList(
         field: 'ModelField',
         config: 'BaseConfig',
     ):
-        from docarray.array.stacked.array_stacked import DocVec
+        from docarray.array.doc_vec.doc_vec import DocVec
 
         if isinstance(value, (cls, DocVec)):
             return value
