+++ docarray/array/doc_vec/list_advance_indexing.py
